requirements.txt
beautifulsoup4==4.12.2
certifi==2022.12.7
charset-normalizer==3.1.0
click==8.1.3
Flask==2.2.3
Flask-Login==0.6.2
Flask-SQLAlchemy==3.0.3
greenlet==2.0.2
idna==3.4
importlib-metadata==6.0.0
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.2
mpmath==1.3.0
numpy==1.24.2
opencv-python==4.6.0.66
pandas==1.5.3
pyserial==3.5
python-dateutil==2.8.2
pytz==2023.2
requests==2.28.2
RPi.GPIO==0.7.1
six==1.16.0
smbus==1.1.post2
soupsieve==2.4
SQLAlchemy==2.0.5.post1
sympy==1.11.1
typing-extensions==4.5.0
urllib3==1.26.15
Werkzeug==2.2.3
zipp==3.15.0
